insert usecase 25

+++ src/main/java/com/napier/sem/Population.java
@@ -6,12 +6,12 @@ public class Population {
     public String name;
 
     //The total population of continent, region or country
-    public int population;
+    public long population;
 
     //The total Population of continent, region or country living in cities (including a %)
-    public int population_c;
+    public long population_c;
 
     //The total Population of continent, region or country not living in cities (including a %)
-    public int population_nc;
+    public long population_nc;
 
 }
